+++ Src/Newtonsoft.Json.UnityConverters.Tests/ConvertingUnityTypes/Physics/Dynamics/JointDriveTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.Physics
+namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.Physics.Dynamics
 {
     public class JointDriveTests : ValueTypeTester<JointDrive>
     {
