+++ Src/Newtonsoft.Json.UnityConverters.Tests/ConvertingUnityTypes/Physics/Dynamics/JointLimitsTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.Physics
+namespace Newtonsoft.Json.UnityConverters.Tests.ConvertingUnityTypes.Physics.Dynamics
 {
     public class JointLimitsTests : ValueTypeTester<JointLimits>
     {
